pyproject.toml
[tool.poetry]
name = "inaina"
version = "0.1.0"
description = ""
authors = ["이한영 <dev@lhy.kr>"]

[tool.poetry.dependencies]
python = "^3.8"
black = "stable"
boto3 = "<2"
django = "^3"
django-dbbackup = "^3"
django-extensions = "<3"
django-quill-editor = "^0.1.22"
django-storages = "<2"
gunicorn = "<21"
pillow = "^8"
psycopg2-binary = "<3"
sentry-sdk = "<2"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "<4"
notebook = "<7"

[tool.black]
line-length = 130

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
